% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_OTRs.R
\name{aiptw_tes}
\alias{aiptw_tes}
\title{Function to estimate treatment effects using AIPTW}
\usage{
aiptw_tes(
  df,
  d_pred,
  Y_name,
  A_name,
  W_list,
  Z_list,
  nuisance_models,
  sl.library.outcome,
  sl.library.treatment,
  sl.library.missingness,
  k_folds,
  ps_trunc_level,
  outcome_type
)
}
\description{
Function to estimate treatment effects using AIPTW
}
