% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_OTR.R
\name{apply_OTR}
\alias{apply_OTR}
\title{Function to apply existing OTR(s) to external data}
\usage{
apply_OTR(
  df,
  CATE_models,
  Y_name,
  A_name,
  W_list,
  Z_list,
  id_name = NULL,
  threshold = c("0.05"),
  te_method = "aiptw",
  nuisance_models = NULL,
  sl.library.outcome = NULL,
  sl.library.treatment = NULL,
  sl.library.missingness = NULL,
  k_folds = 5,
  ps_trunc_level = 0.01,
  outcome_type = "gaussian",
  truncate_CATE = "TRUE"
)
}
\arguments{
\item{df}{dataframe containing external dataset to apply rule(s) to; must contain Z_list variables that are same as pre-trained rule(s)}

\item{CATE_models}{list of CATE model(s) to apply to external dataset}

\item{Y_name}{name of outcome variable. Outcome variable in rule being applied should be the same as outcome variable in new data.}

\item{A_name}{name of treatment variable. Treatment variable in rule being applied should be the same as the treatment variable in the new data.}

\item{W_list}{character vector containing names of covariates in the dataframe used for nuisance models.}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule; must be same names as used in pre-fit CATE model(s))}

\item{id_name}{name of participant ID variable}

\item{threshold}{character vector of decision thresholds for CATE to determine OTR. Values should be positive if `Y_name` is desirable outcome, negative if `Y_name` is undesirable outcome. If threshold is 0, use +0 for desirable, -0 for undesirable.}

\item{nuisance_models}{list of objects of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models (only include if using pre-fit nuisance models)}

\item{sl.library.outcome}{character vector of SuperLearner libraries to use to fit the outcome models}

\item{sl.library.treatment}{character vector of SuperLearner libraries to use to fit the treatment models}

\item{sl.library.missingness}{character vector of SuperLearner libraries to use to fit the missingness models}

\item{k_folds}{integer number of folds to use for nuisance model cross-validation (must specify if fitting nuisance models here. Otherwise uses k from `k_fold_assign_and_CATE`)}

\item{ps_trunc_level}{numeric level below which propensity scores will be truncated (to avoid errors in computing AIPTW)}

\item{outcome_type}{outcome_type specifying continuous (outcome_type = "gaussian") or binary (outcome_type = "binomial") outcome Y (if not providing pre-fit nuisance models)}

\item{truncate_CATE}{logical to indicate if large CATE predictions should be truncated at -1 and 1 (default = TRUE)}

\item{k_fold_assign_and_CATE}{dataframe containing ids, fold assignments, and CATE estimate for each observation in df (only include if using pre-fit nuisance models)}

\item{validRows}{list containing validRows assignents for nuisance models using all observations (only include if using pre-fit nuisance models)}
}
\value{
\describe{
 \item{\code{results}}{list of `Results` objects for each threshold. See description of `Results` object in `compute_estimates`}
 \item{\code{nuisance_models}}{list of `Nuisance` objects containing outcome, treatment, and missingness models used in each fold}
 \item{\code{CATE_models}}{CATE model used in each fold}
 \item{\code{Z_list}}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule)}
 }
}
\description{
Function to apply existing OTR(s) to external data
}
