% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_across_seeds}
\alias{average_across_seeds}
\title{Helper function to average results across multiple seeds}
\usage{
average_across_seeds(results_list, threshold)
}
\arguments{
\item{results_list}{list of results using same models & Z_list, different seeds}

\item{threshold}{threshold of results to use for comparison}
}
\value{
summary_df dataframe summarizing results across length(results_list) seeds
}
\description{
Helper function to average results across multiple seeds
}
