% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{calc_aiptw}
\alias{calc_aiptw}
\title{Function to calculate AIPTW}
\usage{
calc_aiptw(
  a,
  A,
  A_name,
  W,
  Y,
  I_Y,
  d_pred,
  mean_dZ,
  outcome_model,
  treatment_model,
  missingness_model,
  ps_trunc_level,
  idx_sub
)
}
\arguments{
\item{a}{value of treatment to set for all observations}

\item{A}{vector of observed treatments}

\item{A_name}{name of treatment variable}

\item{W}{dataframe of covariates}

\item{Y}{vector of observed outcome variable}

\item{I_Y}{indicator that Y is missing (I_Y = 1 if missing, I_Y = 0 if observed)}

\item{d_pred}{vector of treatment decisions (d_pred = 1 if recommend treatment under rule)}

\item{mean_dZ}{proportion recommended treatment under rule}

\item{outcome_model}{outcome Nuisance model}

\item{treatment_model}{treatment Nuisance model}

\item{missingness_model}{missingness Nuisance model}

\item{ps_trunc_level}{threshold to truncate propensity score}

\item{idx_sub}{indices of observations that are recommended treatment (d_pred == 1)}
}
\value{
\describe{
        (1) a - value of treatment
        (2) aiptw - vector aiptw estimators
        (3) plug_in_est - vector of plug-in estimators
        (4) augmentation - vector of augmentation terms
        (5) var_aug - variance of augmentation term
        }
}
\description{
Function to calculate AIPTW
}
