% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare.otr_results}
\alias{compare.otr_results}
\title{Compare outcomes under different treatment rules}
\usage{
compare.otr_results(
  res_rule1,
  res_rule2,
  threshold1,
  threshold2,
  rule1_comp,
  rule2_comp,
  ...
)
}
\arguments{
\item{res_rule1}{An \code{"otr_results"} object for treatment rule 1}

\item{res_rule2}{An \code{"otr_results"} object for treatment rule 2}

\item{threshold1}{threshold to use for treatment rule 1}

\item{threshold2}{threshold to use for treatment rule 2}

\item{rule1_comp}{Effect type for rule 1 to use in comparison ("treatment effect"/"te" or "subgroup effect"/"se")}

\item{rule2_comp}{Effect type for rule 2 to use in comparison ("treatment effect"/"te" or "subgroup effect"/"se")}

\item{...}{Other arguments (not used)}
}
\value{
dataframe containing expected value and variance for comparison
}
\description{
Compare outcomes under different treatment rules
}
