% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_estimates}
\alias{compute_estimates}
\title{Make treatment decisions and compute estimated outcomes/treatment effects}
\usage{
compute_estimates(
  df,
  Y_name,
  A_name,
  W_list,
  Z_list,
  k_fold_assign_and_CATE,
  nuisance_models,
  CATE_models,
  threshold,
  ps_trunc_level = 0.01,
  truncate_CATE
)
}
\arguments{
\item{df}{dataframe containing dataset}

\item{Y_name}{name of outcome variable in df}

\item{A_name}{name of treatment variable in df}

\item{W_list}{character vector containing names of covariates in the dataframe to be used in nuisance models}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule)}

\item{k_fold_assign_and_CATE}{dataframe containing ids, fold assignments, and CATE estimate for each observation in df}

\item{nuisance_models}{list of objects of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models}

\item{CATE_models}{list of discrete SuperLearner models for CATE from each fold}

\item{threshold}{character vector of decision thresholds for CATE to determine OTR. Values should be positive if `Y_name` is desirable outcome, negative if `Y_name` is undesirable outcome. If threshold is 0, use +0 for desirable, -0 for undesirable.}

\item{ps_trunc_level}{numeric level to use for truncation of any predicted values that fall below it}

\item{truncate_CATE}{logical to indicate if large CATE predictions should be truncated at -1 and 1 (default = TRUE)}
}
\value{
\describe{
 List of objects of class `Results`. Each object contains the following for a given threshold:
 \item{\code{aggregated_results}}{dataframe of overall results aggregated across `k` folds for given threshold}
 \item{\code{k_fold_results}}{list of results by fold for given threshold}
 \item{\code{decision_df}}{original dataset with decision made for each observation at a given threshold}
 \item{\code{k_non_na}}{folds that did not have proportion treated = 1 or = 0 (causing some effect estimates to be NA)}
 }
}
\description{
Make treatment decisions and compute estimated outcomes/treatment effects
}
