% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_OTRs.R
\name{compute_estimates_external}
\alias{compute_estimates_external}
\title{Function to compute estimates on external dataset}
\usage{
compute_estimates_external(
  df,
  Y_name,
  A_name,
  W_list,
  Z_list,
  CATE_preds,
  threshold,
  te_method,
  nuisance_models,
  sl.library.outcome,
  sl.library.treatment,
  sl.library.missingness,
  k_folds,
  ps_trunc_level,
  outcome_type
)
}
\arguments{
\item{df}{dataframe containing external dataset to apply rule(s) to; must contain Z_list variables that are same as pre-trained rule(s)}

\item{Y_name}{name of outcome variable in df}

\item{A_name}{name of treatment variable in df}

\item{W_list}{character vector containing names of covariates in the dataframe to be used for fitting nuisance models}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule; must be same names as used in pre-fit CATE model(s))}

\item{CATE_preds}{dataframe with CATE predictions for each CATE model & truncation flags for extreme predictions}

\item{threshold}{character vector of decision thresholds for CATE to determine OTR. Values should be positive if `Y_name` is desirable outcome, negative if `Y_name` is undesirable outcome. If threshold is 0, use +0 for desirable, -0 for undesirable.}

\item{te_method}{method for estimating treatment effects, default aiptw}

\item{nuisance_models}{list of objects of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models}
}
\value{
tbd
}
\description{
Function to compute estimates on external dataset
}
