% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_CATE.R
\name{learn_CATE_k}
\alias{learn_CATE_k}
\title{Function to learn model for CATE in kth fold}
\usage{
learn_CATE_k(df, Z_list, sl.library.CATE)
}
\arguments{
\item{df}{dataframe containing training dataset}

\item{Z_list}{character vector containing names of variables in df to use to fit CATE model (variables used in treatment rule)}

\item{sl.library.CATE}{character vector of SuperLearner libraries to use to fit the CATE model}
}
\value{
SuperLearner model for CATE
}
\description{
Function to learn model for CATE in kth fold
}
\keyword{internal}
