% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_nuisance.R
\name{learn_nuisance_k}
\alias{learn_nuisance_k}
\title{Estimate nuisance models (outcome, treatment, and missingness) and calculate CATE hats for kth fold}
\usage{
learn_nuisance_k(
  df,
  Y_name,
  A_name,
  W_list,
  sl.library.outcome,
  sl.library.treatment,
  sl.library.missingness,
  outcome_type,
  ps_trunc_level = 0.01
)
}
\arguments{
\item{df}{dataframe containing training dataset}

\item{Y_name}{name of outcome variable in df}

\item{A_name}{name of treatment variable in df}

\item{W_list}{character vector containing names of covariates in the dataframe to be used for fitting nuisance models}

\item{sl.library.outcome}{character vector of SuperLearner libraries to use to fit the outcome model}

\item{sl.library.treatment}{character vector of SuperLearner libraries to use to fit the treatment model}

\item{sl.library.missingness}{character vector of SuperLearner libraries to use to fit the missingness model}

\item{outcome_type}{specifying continuous (gaussian) or binomial outcome Y}

\item{ps_trunc_level}{numeric level to use for truncation of any predicted values that fall below it}
}
\value{
\describe{
 \item{\code{k_fold_nuisance}}{object of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models}
 \item{\code{k_fold_assign_and_CATE}}{numeric vector of CATE estimates for data in df}
 }
}
\description{
Note training dataset contains all observations EXCEPT those assigned to kth fold
}
\keyword{internal}
