% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_OTR.R
\name{predict_CATE_external}
\alias{predict_CATE_external}
\title{Function to predict CATEs & get treatment decisions on external dataset}
\usage{
predict_CATE_external(df, CATE_models, Z_list, truncate_CATE = TRUE)
}
\arguments{
\item{df}{dataframe containing external dataset to apply rule(s) to; must contain Z_list variables that are same as pre-trained rule(s)}

\item{CATE_models}{list of CATE model(s) to apply to external dataset}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule; must be same names as used in pre-fit CATE model(s))}

\item{truncate_CATE}{logical to indicate if large CATE predictions should be truncated at -1 and 1 (default = TRUE)}
}
\value{
dataframe with columns for id, CATEs for each model, and average CATE for rule
}
\description{
Function to predict CATEs & get treatment decisions on external dataset
}
