% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.average_results}
\alias{print.average_results}
\title{Print the output of a \code{"average_results"} object.}
\usage{
\method{print}{average_results}(x, ...)
}
\arguments{
\item{x}{An \code{"otr_results"} or \code{"full_otr_results"} object.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"average_results"} object.
}
