% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.full_otr_results}
\alias{print.full_otr_results}
\title{Print the output of a \code{"full_otr_results"} object.}
\usage{
\method{print}{full_otr_results}(x, ...)
}
\arguments{
\item{x}{A \code{"full_otr_results"} object.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"full_otr_results"} object.
}
