% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.otr_comparison}
\alias{print.otr_comparison}
\title{Print the results of otr_comparison}
\usage{
\method{print}{otr_comparison}(x, ...)
}
\arguments{
\item{x}{An \code{"otr_comparison"} object}

\item{...}{other arguments (not used)}
}
\description{
Print the results of otr_comparison
}
