% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.otr_results}
\alias{print.otr_results}
\title{Print the output of an \code{"otr_results"} object.}
\usage{
\method{print}{otr_results}(x, ...)
}
\arguments{
\item{x}{An \code{"otr_results"} object.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of an \code{"otr_results"} object.
}
