% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_cate}
\alias{strip_cate}
\title{Helper function to apply strip_glm function to libraries in CATE model}
\usage{
strip_cate(cate_model)
}
\arguments{
\item{cate_model}{CATE model fit by SuperLearner}
}
\value{
cate_model of reduced output size
}
\description{
Helper function to apply strip_glm function to libraries in CATE model
}
