% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_nuisance}
\alias{strip_nuisance}
\title{Helper function to apply strip_glm function to libraries in Nuisance model}
\usage{
strip_nuisance(nuisance_model)
}
\arguments{
\item{nuisance_model}{Nuisance model fit by SuperLearner}
}
\value{
Nuisance model of reduced output size
}
\description{
Helper function to apply strip_glm function to libraries in Nuisance model
}
