% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_estimate_k}
\alias{compute_estimate_k}
\title{Make treatment decisions and compute estimated outcomes/treatment effects in kth fold}
\usage{
compute_estimate_k(
  df,
  Y_name,
  A_name,
  W_list,
  Z_list,
  CATE_model,
  nuisance,
  threshold,
  ps_trunc_level = 0.01
)
}
\arguments{
\item{df}{dataframe containing testing dataset}

\item{Y_name}{name of outcome variable in df}

\item{A_name}{name of treatment variable in df}

\item{W_list}{character vector containing names of covariates in the dataframe to be used in nuisance models}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule)}

\item{CATE_model}{discrete SuperLearner model for CATE}

\item{nuisance}{object of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models}

\item{threshold}{decision threshold for CATE to determine OTR. `treatment` should be positive if `Y_name` is desirable outcome, negative if `Y_name` is undesirable outcome}

\item{ps_trunc_level}{numeric level to use for truncation of any predicted values that fall below it}
}
\value{
a list with (1) dataframe with estimated treatment effect among optimally treated, (2) dataframe with estimated outcome if not treated among those who should be treated by decision rule, (3) overall treatment effect among optimally treated, (4) influence function matrix, (5) original kth fold data with corresponding treatment decisions
}
\description{
Make treatment decisions and compute estimated outcomes/treatment effects in kth fold
}
\keyword{internal}
