% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_estimates}
\alias{compute_estimates}
\title{Make treatment decisions and compute estimated outcomes/treatment effects}
\usage{
compute_estimates(
  df,
  Y_name,
  A_name,
  W_list,
  Z_list,
  k_fold_assign_and_CATE,
  nuisance_models,
  CATE_models,
  threshold,
  ps_trunc_level = 0.01
)
}
\arguments{
\item{df}{dataframe containing dataset}

\item{Y_name}{name of outcome variable in df}

\item{A_name}{name of treatment variable in df}

\item{W_list}{character vector containing names of covariates in the dataframe to be used in nuisance models}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule)}

\item{k_fold_assign_and_CATE}{dataframe containing pids, fold assignments, and CATE estimate for each observation in df}

\item{nuisance_models}{list of objects of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models}

\item{CATE_models}{list of discrete SuperLearner models for CATE from each fold}

\item{threshold}{decision threshold for CATE to determine OTR. `treatment` should be positive if `Y_name` is desirable outcome, negative if `Y_name` is undesirable outcome}

\item{ps_trunc_level}{numeric level to use for truncation of any predicted values that fall below it}
}
\value{
\describe{
 \item{\code{overall_results}}{dataframe of overall results aggregated across `k` folds}
 \item{\code{EY_A1_d1}}{dataframe of AIPTW for optimally treated in each fold}
 \item{\code{EY_A0_d1}}{dataframe of AIPTW for not treating those who should be treated under decision rule in each fold}
 \item{\code{treatment_effect}}{dataframe of treatment effect in each fold}
 \item{\code{influence_fns}}{influence functions used in each fold}
 \item{\code{decision_df}}{original dataset with decision made for each observation}
 }
}
\description{
Make treatment decisions and compute estimated outcomes/treatment effects
}
