% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_OTR.R
\name{compute_estimates_external}
\alias{compute_estimates_external}
\title{Function to compute estimates on external dataset}
\usage{
compute_estimates_external(
  df,
  Y_name,
  A_name,
  W_list,
  Z_list,
  CATE_preds,
  threshold,
  nuisance_models,
  sl.library.outcome,
  sl.library.treatment,
  sl.library.missingness,
  k_folds,
  ps_trunc_level,
  outcome_type,
  id_name = NULL
)
}
\arguments{
\item{df}{dataframe containing external dataset to apply rule(s) to; must contain Z_list variables that are same as pre-trained rule(s)}

\item{Y_name}{name of outcome variable in df}

\item{A_name}{name of treatment variable in df}

\item{W_list}{character vector containing names of covariates in the dataframe to be used for fitting nuisance models}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule; must be same names as used in pre-fit CATE model(s))}

\item{CATE_preds}{dataframe with CATE predictions for each CATE model & truncation flags for extreme predictions}

\item{threshold}{character vector of decision thresholds for CATE to determine OTR. Values should be positive if `Y_name` is desirable outcome, negative if `Y_name` is undesirable outcome. If threshold is 0, use +0 for desirable, -0 for undesirable.}

\item{nuisance_models}{list of objects of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models}

\item{te_method}{method for estimating treatment effects, default aiptw}
}
\value{
\describe{
List of "Results" objects for each threshold. Each results object contains:
\item{\code{aggregated_results}}{Results averaged over each nuisance model}
\item{\code{k_fold_results}}{Results for each nuisance model}
\item{\code{decision_df}}{Dataframe with CATE prediction from every individual CATE_model, flag for CATE truncation, average CATE predition used for treatment decision, treatment decision}
\item{\code{k_non_na}}{equal to number of k_folds used for nuisance models. Ignore-- placeholder to keep same format as Results object from internal compute estimates}
}
}
\description{
Function to compute estimates on external dataset
}
