% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_OTR.R
\name{estimate_OTR}
\alias{estimate_OTR}
\title{Main function to calculate estimated treatment effects for treatment rule}
\usage{
estimate_OTR(
  df,
  Y_name,
  A_name,
  W_list,
  Z_list,
  id_name = NULL,
  sl.library.CATE,
  nuisance_models = NULL,
  k_fold_assign_and_CATE = NULL,
  sl.library.outcome = NULL,
  sl.library.treatment = NULL,
  sl.library.missingness = NULL,
  threshold = 0.05,
  k_folds = 2,
  ps_trunc_level = 0.01,
  outcome_type = "gaussian"
)
}
\arguments{
\item{df}{dataframe containing full dataset}

\item{Y_name}{name of outcome variable in df}

\item{A_name}{name of treatment variable in df}

\item{W_list}{character vector containing names of covariates in the dataframe to be used for fitting nuisance models}

\item{Z_list}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule)}

\item{sl.library.CATE}{character vector of SuperLearner libraries to use to fit the CATE model}

\item{nuisance_models}{list of objects of class `Nuisance` containing outcome, treatment, and missingness SuperLearner models (only include if using pre-fit nuisance models)}

\item{k_fold_assign_and_CATE}{dataframe containing ids, fold assignments, and CATE estimate for each observation in df (only include if using pre-fit nuisance models)}

\item{sl.library.outcome}{character vector of SuperLearner libraries to use to fit the outcome models}

\item{sl.library.treatment}{character vector of SuperLearner libraries to use to fit the treatment models}

\item{sl.library.missingness}{character vector of SuperLearner libraries to use to fit the missingness models}

\item{threshold}{decision threshold for CATE to determine OTR. `treatment` should be positive if `Y_name` is desirable outcome, negative if `Y_name` is undesirable outcome}

\item{k_folds}{integer number of folds to use for cross-validation (must specify if fitting outcome, treatment, and missingness models. Otherwise uses k from `k_fold_assign_and_CATE`)}

\item{ps_trunc_level}{numeric evel below which propensity scores will be truncated (to avoid errors in computing AIPTW)}

\item{outcome_type}{outcome_type specifying continuous (outcome_type = "gaussian") or binary (outcome_type = "binomial") outcome Y (if not providing pre-fit nuisance models)}
}
\value{
\describe{
 \item{\code{overall_results}}{dataframe of overall results aggregated across `k` folds}
 \item{\code{EY_A1_d1}}{dataframe of AIPTW for optimally treated in each fold}
 \item{\code{EY_A0_d1}}{dataframe of AIPTW for not treating those who should be treated under decision rule in each fold}
 \item{\code{treatment_effect}}{dataframe of treatment effect in each fold}
 \item{\code{decision_df}}{original dataset with decision made for each observation}
 \item{\code{CATE_models}}{CATE model used in each fold}
 \item{\code{Z_list}}{character vector containing names of variables in df used to fit CATE model (variables used in treatment rule)}
 }
}
\description{
Primary function to estimate treatment effects for a given treatment rule.
Can fit nuisance models internally or be provided with pre-fit nuisance models for the given dataset.
}
