% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_CATE.R
\name{learn_CATE}
\alias{learn_CATE}
\title{Function to learn model for CATE using k-fold cross validation}
\usage{
learn_CATE(df, Z_list, k_fold_assign_and_CATE, sl.library.CATE, validRowsAll)
}
\arguments{
\item{df}{dataframe containing full dataset}

\item{Z_list}{character vector containing names of variables in df to use to fit CATE model (variables used in treatment rule)}

\item{k_fold_assign_and_CATE}{dataframe containing ids, fold assignments, and CATE estimate for each observation in df}

\item{sl.library.CATE}{character vector of SuperLearner libraries to use to fit the CATE model}

\item{validRowsAll}{validRows SuperLearner row assignments from nuisance models using all rows in data (missingness, treatment models)}
}
\value{
list containing SuperLearner model for CATE in each fold
}
\description{
Function to learn model for CATE using k-fold cross validation
}
